compilable !

+++ firmware/include/Robot.h
@@ -1,13 +1,15 @@
 #ifndef ROBOT_H_
 #define ROBOT_H_
 
+#include <stdint.h>
+
 #ifdef __cplusplus
 extern "C" {
 #endif
 
 bool robot_init(void);
 void robot_connect(void);
-void robot_save_configuration(int configuration, int speed, float drift);
+void robot_save_configuration(int configuration, int speed, uint8_t drift);
 void robot_flip(void);
 int robot_serialize_for_request(char* buffer);
